Upgrade iiif module
This commit includes some changes to get ndnp_iiif working with the latest
[iiif](http://github.com/zimeon/iiif) module. There were some backward
incompatible changes in the static image tile generation. Eventhough we
are likely going to be using a separate IIIF image server I thought it would be
useful to get it working again.

Running the test suite will generate IIIF data and place a demo Mirador
viewer alongside which you can use to verify that the newspaper manifest
and tiles are functional.

    cd test-data/iiif
    python -m SimpleHTTPServer

then open http://localhost:8000