separate modules
The reading of NDNP data into a set of objects was getting a bit tangled up with
the writing of those objects as IIIF data. This was particularly apparent when
it came to thinking about how to make the generation of thumbnails optional.

This commit includes a substantial refactor to separate the reading of NDNP
data into ndnp_iiif.reader and the writing of IIIF data into ndnp_iiif.writer.